+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.2 (November 24, 2022)
 
-- Fixed `onSetError` not receiving the correct `setItems` object (#15)
+- Fixed `onSetError` not receiving the correct `setItems` object ([#15](https://github.com/PixievoltNo1/svelte-webext-storage-adapter/issues/15))
 
 ## 2.0.1 (November 20, 2021)